refactor(product): split transcript selection mechanics
